refactoring components and slices code structure
