+++ src/components/Movies/movies.scss
