+++ src/components/index.js
@@ -1,0 +1,19 @@
+import Header from "./Header/Header";
+import Modal from "./Modal/Modal";
+import Movie from "./Movie/Movie";
+import Movies from "./Movies/Movies";
+import ScrollContent from "./ScrollContent/ScollContent";
+import Starred from "./Starred/Starred";
+import WatchLater from "./WatchLater/WatchLater";
+import YoutubePlayer from "./YoutubePlayer/YoutubePlayer";
+
+export {
+  Header,
+  Modal,
+  Movie,
+  Movies,
+  ScrollContent,
+  Starred,
+  WatchLater,
+  YoutubePlayer,
+};
