+++ src/store/watchLaterSlice.js
@@ -13,7 +13,7 @@ const watchLaterSlice = createSlice({
             const indexOfId = state.watchLaterMovies.findIndex(key => key.id === action.payload.id)
             state.watchLaterMovies.splice(indexOfId, 1)
         },
-        remveAllWatchLater: (state) => {
+        removeAllWatchLater: (state) => {
             state.watchLaterMovies = []
         },
     },
